Renamed the semantics lib to analyzer.
