+++ analyzer/src/namespace/events.rs
