+++ analyzer/src/namespace/mod.rs
