+++ analyzer/src/namespace/operations.rs
