+++ analyzer/src/namespace/scopes.rs
