+++ analyzer/src/namespace/types.rs
