+++ analyzer/src/traversal/_utils.rs
