+++ analyzer/src/traversal/assignments.rs
