+++ analyzer/src/traversal/contracts.rs
