+++ analyzer/src/traversal/declarations.rs
