+++ analyzer/src/traversal/functions.rs
