+++ analyzer/src/traversal/mod.rs
