+++ analyzer/src/traversal/module.rs
