+++ analyzer/src/traversal/types.rs
