+++ compiler/Cargo.toml
@@ -14,7 +14,7 @@ solc-backend = ["solc"]
 [dependencies]
 fe-common = {path = "../common", version = "^0.1.0-alpha"}
 fe-parser = {path = "../parser", version = "^0.1.0-alpha"}
-fe-semantics = {path = "../semantics", version = "^0.1.0-alpha"}
+fe-analyzer = {path = "../analyzer", version = "^0.1.0-alpha"}
 serde_json = "1.0"
 serde = "1.0"
 hex = "0.4"
