+++ compiler/src/yul/operations/abi.rs
@@ -1,7 +1,7 @@
 use crate::yul::names;
 use crate::yul::operations::data as data_operations;
 use crate::yul::utils;
-use fe_semantics::namespace::types::{
+use fe_analyzer::namespace::types::{
     AbiArraySize,
     AbiDecodeLocation,
     AbiEncoding,
@@ -82,7 +82,7 @@ mod tests {
         encode,
         encode_size,
     };
-    use fe_semantics::namespace::types::{
+    use fe_analyzer::namespace::types::{
         AbiDecodeLocation,
         FeString,
         U256,
